+++ src/mongocxx/include/mongocxx/v_noabi/mongocxx/structured_log-fwd.hpp
@@ -19,16 +19,6 @@
 #include <mongocxx/config/prelude.hpp>
 
 namespace mongocxx {
-namespace v_noabi {
-
-using v1::structured_log_component;
-using v1::structured_log_entry;
-using v1::structured_log_level;
-
-} // namespace v_noabi
-} // namespace mongocxx
-
-namespace mongocxx {
 
 using v1::structured_log_component;
 using v1::structured_log_entry;