+++ src/mongocxx/include/mongocxx/v_noabi/mongocxx/structured_logging.hpp
@@ -14,37 +14,15 @@
 
 #pragma once
 
-#include <mongocxx/options/structured_logging-fwd.hpp> // IWYU pragma: export
+#include <mongocxx/structured_logging-fwd.hpp> // IWYU pragma: export
 
 //
 
 #include <mongocxx/v1/structured_logging.hpp> // IWYU pragma: export
 
-#include <mongocxx/config/prelude.hpp>
-
-namespace mongocxx {
-namespace v_noabi {
-namespace options {
-
-using v1::structured_logging;
-
-} // namespace options
-} // namespace v_noabi
-} // namespace mongocxx
-
-namespace mongocxx {
-namespace options {
-
-using v_noabi::options::structured_logging;
-
-} // namespace options
-} // namespace mongocxx
-
-#include <mongocxx/config/postlude.hpp>
-
 ///
 /// @file
-/// Provides @ref mongocxx::v_noabi::options::structured_logging.
+/// Provides @ref mongocxx::structured_logging.
 ///
 /// @par Includes
 /// - @ref mongocxx/v1/structured_logging.hpp